+++ tomcat/src/test/java/org/apache/coyote/http11/NotFoundControllerTest.java
@@ -1,4 +1,4 @@
-package com.techcourse.controller;
+package org.apache.coyote.http11;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -6,9 +6,9 @@ import java.io.File;
 import java.net.URL;
 import java.nio.file.Files;
 import java.util.List;
-import org.apache.NotFoundController;
 import org.apache.coyote.HttpRequest;
 import org.apache.coyote.HttpResponse;
+import org.apache.coyote.NotFoundController;
 import org.apache.coyote.RequestHandler;
 import org.apache.coyote.http11.request.Http11Request;
 import org.apache.coyote.http11.request.Http11RequestBody;
